Evalueer samen de resultaten van 1 of meer steekproeven op uitgaand geld

Het samennemen van de resultaten gebeurt door convolutie van de foutkanskrommes van de afzonderlijke steekproeven tot 1 foutkanskromme.

Waar van toepassing worden 100%-getoetste posten (het hoogstratum of topstratum) als deterministische factoren opgeteld bij de resulterende statistische verdeling.

We berekenen de meest waarschijnlijke en de maximale fout als fractie en in geld.

De meest waarschijnlijke fout is de modus van de kanskromme. De maximale fout is afhankelijk van de gevraagde zekerheid, en is de fout bij een cumulatieve kans gelijk aan deze zekerheid.

De statistische interpretatie van de risico waarden hoog, midden en laag voor IHR, IBR en CAR, die deze module hanteert is volgens het HARO, het Handboek Auditing Rijksoverheid. Het HARo wordt beheerd door Auditdienst Rijk, de ADR.

Usage

eval_stratified(
  steekproeven,
  model = "binomiaal",
  zekerheid = 0.95,
  MC = 1e+07,
  start = 1,
  vergelijk = TRUE
)

Arguments

steekproeven

Een tibble. Elke regel van de tibble beschrijft 1 steekproef, dus 1 van de genomen steekproeven. De tibble eist de volgende expliciete kolommen voor redundantie en veiligheid: naam, een aanduiding van de steekproef, waarde_laag, de omvang in geld van de massa waaruit de steekproef (laagstratum) is getrokken, n_laag, het aantal getrokken posten in het laagstratum, k_laag, de som van de foutfracties van de posten in het laagstratum, ihr, inherent risico, te weten H, M of L, ibr, intern beheersingsrisico, te weten H, M of L, car, cijferanalyserisico, te weten H, M of L, en materialiteit, als fractie van de totale massa. fout_hoog, Het gevonden foutbedrag in het 100%-getoetste topstratum. goed_hoog, Het goedgekeurde bedrag in het 100%-getoetste topstratum. n_hoog, Het aantal posten in het hoogstratum. n_totaal, Het totale aantal posten in deze audit (n_laag + n_hoog). waarde_hoog, De totale boekwaarde van het hoogstratum (fout_hoog + goed_hoog). waarde_populatie, De totale boekwaarde van de hele populatie (waarde_laag + waarde_hoog).

model

Het statistische model dat gebruikt wordt voor de extrapolatie. Keuze uit "binomiaal" (standaard) of "poisson". #' @param zekerheid Het zekerheidsniveau waarop we de maximale foutfractie berekenen.

MC

Het aantal Monte Carlo iteraties dat gebruikt wordt. Monte Carlo berekeningen baseren zich op toevalsgetallen.

start

Startwaarde voor de toevalsgenerator.

vergelijk

TRUE of FALSE, als TRUE dan worden wat vergelijkende berekeningen uitgevoerd en de resultaten daarvan toegevoegd aan de uitkomst van de functie.

Value

Een lijst, bestaande uit de convolutie-uitkomsten (fracties en geld), eventuele vergelijkingen, en de verrijkte invoergegevens.

Details

We gaan uit van de som van de foutfracties, de k-waarde, dus we kijken niet naar de foutfracties per post.

De maximale fout wordt bepaald aan de hand van de resulterende kanskromme, op basis van de gewenste zekerheid. Visueel is de maximale fout, pm, te bepalen in een tweedimensionaal, haaks, assenstelsel. De horizontale as, de p-as, loopt van 0 tot 1. De waarden langs die as geven de mogeljke foutfracties weer, lopend van 0 (geen fouten) tot 1 (alles fout). De verticale as, de c-as, loopt van 0 to oneindig. Deze as geeft de kanswaarden van de foutfracties aan. In dit assenstelsel kunnen we de kanskromme afbeelden. Het oppervlak onder de kanskromme is 1. Hierbij praten we over het oppervlak begrenst door de p-as, aan de onderkant, en de verticale lijnen p = 0, en p = 1. pm is het punt op de p-as waarbij de verticale lijn p = pm, het oppervlak onder de kanskromme begrenst zodat links van deze lijn het oppervlak gelijk is aan de zekerheid, bijvoorbeeld 0,95.

Aggregatie is puur op statistische gronden: namelijk risico's op fouten boven de meest waarschijnlijke fout en op onder de meest waarschijnlijke fout vlakken elkaar enigszins uit genomen over de meerdere steekproeven. Dus, bij het aggregeren van de resultaten van de verschillende steekproeven wordt geen enkele aanname gedaan over gelijkenis tussen de eigenschappen van de afzonderlijke administraties waaruit is getrokken.

Deze module kan ook steekproeven combineren over massa's waarvoor een verschillende risicoinschatting geldt.

Bijvoorbeeld: Steekproef1 is gebaseerd op een zekerheid van 95% omdat ihr, ibr en car alledrie op hoog (H) staan. De materialiteit is 2%. Het betreft 100 miljoen euro. Voor steekproef1 trekken we 148 posten, waarbij 1 fout blijkt. Steekproef2 is gebaseerd op een zekerheid van 64% omdat ihr en ibr allebei op laag staan en alleen car op hoog. Het betreft ook 100 miljoen euro en een materialiteit van 2%. Voor steekproef2 trekken we 50 posten waarvan er 0 fout blijken.

Bij een risicoinschatting onder 95% van een of meer van de massa's waarover wordt gestoken worden deze lagere risicoinschattingen vertaald naar extra getrokken foutloze posten. In ons voorbeeld bepalen we voor steekproef2 het aantal foutloze posten nodig om een positieve uitspraak te doen bij 64% en bij 95%. Dit zijn respectievelijk 50 en 148. Het verschil is 148-50 = 98 posten. Daarna berekenen we totale maximale fout op basis van een zekerheid van 95%, en steekproef1, 148 posten waarvan 1 fout, en steekproef2, met 50 + 98 posten, waarvan 0 fout. De maximale fout is dan 1,83% ofwel ongeveer 3.660.000. De meest waarschijnlijke fout 0,52% ofwel ongeveer 1.160.000 euro.

Het is de verantwoordelijkheid van de auditor hoe om te gaan met een steekproef waarbij de risicoinschatting niet op H staat en er toch fouten worden gevonden. Dit probleem staat los van hoe de uitkomsten van meerdere steekproeven samen te nemen.

Als de parameter vergelijk TRUE is doen we, ter vergelijking, ook een evaluatie:

voor elke steekproef los

voor alle steekproeven samen, waarbij ze beschouwd worden als te zijn getrokken op 1 massa, en als 1 steekproef.